Разделил меню и запросы и добавил коментарии что находится в классах
