+++ CLI/Menu.cs
@@ -4,12 +4,9 @@ using DataBaseLib;
 
 namespace CLI
 {
-    public class Menu
+    public class Menu // Этот клас создан для работы с меню
+        //Класс Menu вызывает методы из класса Query
     {
-        const string driverType = "System.Collections.Generic.List`1[DataBaseLib.Driver]";
-        const string CarType = "System.Collections.Generic.List`1[DataBaseLib.Car]";
-        const string ServiceType = "System.Collections.Generic.List`1[DataBaseLib.Service]";
-
         public static bool IsEmpty(string key)
         {
             if (key == string.Empty)
@@ -23,17 +20,17 @@ namespace CLI
 
         static void DriverMenu()
         {
-            Show.SubMenu1();
+            Show.DriverMenu();
             var key = Convert.ToString(Console.ReadLine());
             if (IsEmpty(key)) return;
 
             switch (key)
             {
                 case "1":
-                    PrintDriver();
+                    Query.PrintDriver();
                     break;
                 case "2":
-                    InsertDriver();
+                    Query.InsertDriver();
                     Show.PrintGreen($"Запрос отправлен");
                     break;
                 case "3":
@@ -41,7 +38,6 @@ namespace CLI
                 case "4":
                     break;
                 case "0":
-                    Show.Menu();
                     break;
                 default:
                     Console.WriteLine("Введен некорректный символ");
@@ -54,7 +50,7 @@ namespace CLI
 
         static void CarMenu()
         {
-            Show.SubMenu2();
+            Show.CarMenu();
             var key = Convert.ToString(Console.ReadLine());
             if (IsEmpty(key)) return;
 
@@ -69,7 +65,6 @@ namespace CLI
                 case "4":
                     break;
                 case "0":
-                    Show.Menu();
                     break;
                 default:
                     Console.WriteLine("Введен некорректный символ");
@@ -82,7 +77,7 @@ namespace CLI
 
         static void ServiceMenu()
         {
-            Show.SubMenu3();
+            Show.ServiceMenu();
             var key = Convert.ToString(Console.ReadLine());
             if (IsEmpty(key)) return;
 
@@ -97,7 +92,6 @@ namespace CLI
                 case "4":
                     break;
                 case "0":
-                    Show.Menu();
                     break;
                 default:
                     Console.WriteLine("Введен некорректный символ");
@@ -110,7 +104,7 @@ namespace CLI
 
         static void Info()
         {
-            Show.SubMenu4();
+            Show.Info();
             var key = Convert.ToString(Console.ReadLine());
             if (IsEmpty(key)) return;
 
@@ -121,7 +115,6 @@ namespace CLI
                 case "2":
                     break;
                 case "0":
-                    Show.Menu();
                     break;
                 default:
                     Console.WriteLine("Введен некорректный символ");
@@ -130,56 +123,6 @@ namespace CLI
 
             Console.WriteLine($"Нажмите любую клавишу, чтобы продолжить");
             Console.ReadLine();
-        }
-
-        public static void PrintDriver()
-        {
-            DataBase db = new DataBase();
-            IQuery idriver = new Driver();
-            var list = (List<Driver>)db.Select(idriver);
-            foreach (var driver in list)
-            {
-                Show.PrintRed($"id водителя {driver.Id}");
-                Show.PrintGreen($"{driver.FullName} {driver.Age} лет");
-                Show.PrintYelloy($"Cтаж вождения - {driver.DriveExp}, id автомобиля - {driver.IdCar}\n");
-            }
-        }
-
-        public static void InsertDriver()
-        {
-            var db = new DataBase();
-
-            Show.PrintGreen("Введите Имя и фамилию в одну строчку");
-            var fullName = Convert.ToString(Console.ReadLine());
-
-            Show.PrintGreen("Введите возраст");
-            var age = Convert.ToUInt32(Console.ReadLine());
-
-            Show.PrintGreen("Введите стаж вождения");
-            var drivingExp = Convert.ToUInt32(Console.ReadLine());
-
-            //db.SelectCar();  //  Ждать пока Алексей напишет метод
-            Show.PrintGreen("Введите id автомобиля");
-            var idCar = Convert.ToUInt32(Console.ReadLine());
-
-            // var driver = new Driver
-            // {
-            //     FullName = fullName,
-            //     Age = age,
-            //     DriveExp = drivingExp,
-            //     IdCar = idCar
-            // };
-            // IQuery idriver = driver; 
-            // db.Insert(idriver);
-
-            IQuery idriver = new Driver
-            {
-                FullName = fullName,
-                Age = age,
-                DriveExp = drivingExp,
-                IdCar = idCar
-            };
-            db.Insert(idriver);
         }
 
         public static void SubMenu(string key)
